+++ lib/cli.js
@@ -6,7 +6,6 @@
 "use strict";
 
 let cli = require("commander"),
-    jsdiff = require("diff"),
     fs = require("fs"),
     fsUtils = require("./utils/fs-utils"),
     path = require("path"),
@@ -99,8 +98,8 @@ function lintString(sourceCode, userConfig, errorReporter, fileName) {
                 applyFixes(fileName, result);
                 fixesApplied = result.fixesApplied;
             } else {
-                displayDiff(fileName, sourceCode, result.fixedSourceCode);
-                lintErrors = result.fixesApplied;
+                errorReporter.reportDiff(fileName,
+                    sourceCode, result.fixedSourceCode, result.fixesApplied.length);
             }
         } else {
             lintErrors = solium.lint(sourceCode, userConfig);
@@ -109,7 +108,7 @@ function lintString(sourceCode, userConfig, errorReporter, fileName) {
         // Don't abort in case of a parse error, just report it as a normal lint issue.
         if (e.name !== "SyntaxError") {
             const messageOrStackrace = userConfig.options.debug ? e.stack : e.message;
-            errorReporter.reportFatal(`An error occured while linting over ${fileName}:\n${messageOrStackrace}`);
+            errorReporter.reportFatal(`An error occured while linting over ${fileName}:${EOL}${messageOrStackrace}`);
             process.exit(errorCodes.ERRORS_FOUND);
         }
 
@@ -133,33 +132,6 @@ function lintString(sourceCode, userConfig, errorReporter, fileName) {
 
 function applyFixes(fileName, lintResult) {
     lintResult.fixesApplied.length && fs.writeFileSync(fileName, lintResult.fixedSourceCode);
-}
-
-function displayDiff(fileName, sourceCode, fixedSourceCode) {
-    let diff = jsdiff.structuredPatch(fileName,
-        fileName, sourceCode, fixedSourceCode, "old-header", "new-header");
-
-    if (diff.hunks.length == 0) {
-        return;
-    }
-
-    console.log((`Diff for: ${fileName}`).cyan);
-    diff.hunks.forEach(function(hunk){
-        console.log(`@@ -${hunk.oldStart},${hunk.oldLines} +${hunk.newStart},${hunk.newLines} @@`.cyan);
-        hunk.lines.forEach(function(line) {
-            console.log(getDiffLine(line));
-        });
-    });
-}
-
-function getDiffLine(diffPart) {
-    if (diffPart.startsWith("+")) {
-        return diffPart.green;
-    } else if (diffPart.startsWith("-")) {
-        return diffPart.red;
-    }
-
-    return diffPart;
 }
 
 /**
@@ -250,7 +222,7 @@ function createCliOptions(cliObject) {
         .option("-i, --init", "Create default rule configuration files")
         .option("-f, --file [filepath::String]", "Solidity file to lint")
         .option("-d, --dir [dirpath::String]", "Directory containing Solidity files to lint")
-        .option("-R, --reporter [name::String]", "Format to report lint issues in (pretty | gcc)")
+        .option("-R, --reporter [name::String]", "Format to report lint issues in (pretty | gcc)", "pretty")
         .option("-c, --config [filepath::String]", "Path to the .soliumrc configuration file")
         .option("-, --stdin", "Read input file from stdin")
         .option("--fix", "Fix Lint issues where possible")
@@ -280,8 +252,6 @@ function createCliOptions(cliObject) {
  * @returns {Object} reporter The reporter whose name was supplied.
  */
 function getErrorReporter(name) {
-    name = name || "pretty";
-
     try {
         return require("./reporters/" + name);
     } catch (e) {
@@ -309,7 +279,7 @@ function execute(programArgs) {
     try {
         errorReporter = getErrorReporter(cli.reporter);
     } catch (e) {
-        console.error(`[Fatal error] ${e.message}`);
+        process.stderr.write(`[Fatal error] ${e.message}${EOL}`);
         process.exit(errorCodes.INVALID_PARAMS);
     }
 
@@ -359,8 +329,13 @@ function execute(programArgs) {
         debug: Boolean(cli.debug)
     };
 
-    if (userConfig.options.autofix && userConfig.options.autofixDryrun) {
-        return errorReporter.reportFatal("Cannot use both --fix and --fix-dry-run");
+    if (userConfig.options.autofixDryrun) {
+        if (userConfig.options.autofix) {
+            return errorReporter.reportFatal("Cannot use both --fix and --fix-dry-run");
+        }
+        if (cli.reporter != "pretty") {
+            return errorReporter.reportFatal("Option --fix-dry-run is only supported with pretty reporter");
+        }
     }
 
     userConfig.plugins = userConfig.plugins || [];
