+++ lib/reporters/pretty.js
@@ -5,8 +5,8 @@
 
 "use strict";
 
-const path = require("path"),
-    sort = require("lodash/sortBy"), Table = require("text-table");
+const path = require("path"), jsdiff = require("diff"),
+    sort = require("lodash/sortBy"), Table = require("text-table"), {EOL} = require("os");
 require("colors");
 
 let counts = {};
@@ -18,6 +18,16 @@ function color(type) {
 
 function colorInternalIssue(type) {
     return type === "warning" ? "blue" : "red";
+}
+
+function getDiffLine(diffPart) {
+    if (diffPart.startsWith("+")) {
+        return diffPart.green;
+    } else if (diffPart.startsWith("-")) {
+        return diffPart.red;
+    }
+
+    return diffPart;
 }
 
 
@@ -72,6 +82,22 @@ module.exports = {
         if (Array.isArray(fixesApplied)) {
             counts.fixes = (counts.fixes || 0) + fixesApplied.length;
         }
+    },
+
+    reportDiff(fileName, sourceCode, fixedSourceCode, issuesFixed) {
+        let diff = jsdiff.structuredPatch(fileName,
+            fileName, sourceCode, fixedSourceCode, "old-header", "new-header");
+
+        if (diff.hunks.length == 0) {
+            return;
+        }
+
+        process.stdout.write((`Diff for: ${fileName}${EOL}`).cyan);
+        diff.hunks.forEach(function(hunk){
+            process.stdout.write(`@@ -${hunk.oldStart},${hunk.oldLines} +${hunk.newStart},${hunk.newLines} @@${EOL}`.cyan);
+            hunk.lines.forEach(line => { process.stdout.write(getDiffLine(line) + EOL); });
+        });
+        process.stdout.write(`${EOL}${issuesFixed} lint issue(s) can be fixed.${EOL}`.green);
     },
